pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "meridian-logistics"
version = "0.1.0"
description = "Meridian Logistics Memphis PoC scaffold."
readme = "prd.md"
requires-python = ">=3.10"
dependencies = [
  "alembic>=1.18,<1.19",
  "fastapi[standard]>=0.135,<0.136",
  "jsonschema>=4.25,<5",
  "langgraph>=1.1,<1.2",
  "psycopg[binary]>=3.2,<4",
  "pyyaml>=6.0,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4,<9",
]

[tool.hatch.build.targets.wheel]
packages = ["backend", "evals"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["contracts/tests"]
